+++ formio/controllers/public.py
@@ -50,6 +50,7 @@ class FormioPublicController(http.Controller):
         if form and form.builder_id.schema:
             res['schema'] = json.loads(form.builder_id.schema)
             res['options'] = self._get_public_form_js_options(form)
+            res['locales'] = self._get_public_form_js_locales(form)
             res['params'] = self._get_public_form_js_params(form.builder_id)
 
         args = request.httprequest.args
@@ -142,6 +143,7 @@ class FormioPublicController(http.Controller):
         if formio_builder.schema:
             res['schema'] = json.loads(formio_builder.schema)
             res['options'] = self._get_public_create_form_js_options(formio_builder)
+            res['locales'] = self._get_public_form_js_locales(formio_builder)
             res['params'] = self._get_public_form_js_params(formio_builder)
 
         args = request.httprequest.args
@@ -307,6 +309,9 @@ class FormioPublicController(http.Controller):
             options['language'] = request.env.ref('base.lang_en').formio_ietf_code
 
         return options
+
+    def _get_public_form_js_locales(self, builder):
+        return builder._get_form_js_locales()
 
     def _get_public_form_js_params(self, builder):
         return builder._get_public_form_js_params()