+++ formio/models/formio_builder_translation.py
@@ -1,7 +1,8 @@
 # Copyright Nova Code (http://www.novacode.nl)
 # See LICENSE file for full licensing details.
 
-from odoo import api, fields, models
+from odoo import api, fields, models, _
+from odoo.exceptions import ValidationError
 
 
 class BuilderTranslation(models.Model):
@@ -19,6 +20,25 @@ class BuilderTranslation(models.Model):
     builder_name = fields.Char(related='builder_id.name', string='Builder Name', store=True)
     builder_version = fields.Integer(related='builder_id.version', string='Builder Version', store=True)
     lang_name = fields.Char(related='lang_id.name', string='Language Name', store=True)
+
+    @api.constrains('builder_id', 'lang_id', 'source')
+    def _constraint_unique(self):
+        errors = []
+        for rec in self:
+            res = self.search([
+                ("builder_id", "=", rec.builder_id.id),
+                ("lang_id", "=", rec.lang_id.id),
+                ("source", "=", rec.source)
+            ])
+            if len(res) > 1:
+                errors.append(
+                    _("- Form Builder Name: {name}\n- Source Term: {source}\n- Translation: {translation}").format(
+                        name=rec.builder_id.name, source=rec.source, translation=rec.value
+                    )
+                )
+        if errors:
+            msg = _('Form Builder Translations must be unique.\n\n%s') % '\n\n'.join(errors)
+            raise ValidationError(msg)
 
     @api.depends('lang_id', 'source', 'value')
     def name_get(self):