+++ formio/models/formio_default_asset_css.py
@@ -1,10 +1,8 @@
 # Copyright Nova Code (http://www.novacode.nl)
 # See LICENSE file for full licensing details.
 
-from odoo import api, fields, models, _
+from odoo import api, fields, models
 
-import logging
-logger = logging.getLogger(__name__)
 
 class DefaultAssetCss(models.Model):
     _name = 'formio.default.asset.css'